Fix Generate Training plan
