+++ backend/services/planes/app/models.py
@@ -1,5 +1,4 @@
-import app.config as config
-from sqlmodel import Field, SQLModel, create_engine, Session
+from sqlmodel import Field, SQLModel
 
 
 class Plan(SQLModel, table=True):
@@ -8,13 +7,3 @@ class Plan(SQLModel, table=True):
     nombre: str
     numeroEntrenamientosSemana: str
     objetivoDistanciaEntrenamiento: str
-
-
-SQLALCHEMY_DATABASE_URL = f"postgresql://{config.DB_USER}:{config.DB_PASSWORD}@{config.DB_HOST}:{config.DB_PORT}/{config.DB_NAME}"
-engine = create_engine(SQLALCHEMY_DATABASE_URL)
-SQLModel.metadata.create_all(engine)
-
-
-def create_session():
-    session = Session(engine)
-    return session
