+++ frontWebApp/src/app/pages/generar-plan-entre/plan-entrenamiento.service.ts
@@ -1,20 +1,25 @@
-import {HttpClient} from '@angular/common/http';
+import {HttpClient, HttpHeaders} from '@angular/common/http';
 import {Injectable} from '@angular/core';
 import {Observable} from 'rxjs';
 import {PlanEntrenamiento} from "./plan-entrenamiento";
+import {environment} from "../../../environments/environment";
 
 @Injectable({
   providedIn: 'root'
 })
 export class PlanEntrenamientoService {
 
-  private apiUrl: string = 'http://localhost:4200/generar_plan_entrenamiento';
+  private apiUrl: string = environment.backendUrl;
 
   constructor(private http: HttpClient) {
   }
 
-  generarPlanEntrenamiento(planEntrenamiento: PlanEntrenamiento): Observable<PlanEntrenamiento> {
-    let postPlanEntrenamiento = this.apiUrl + 'planEntrenamiento';
-    return this.http.post<PlanEntrenamiento>(postPlanEntrenamiento, planEntrenamiento);
+  generarPlanEntrenamiento(planEntrenamiento: any): Observable<any> {
+    let postPlanEntrenamiento = this.apiUrl + 'generate';
+    return this.http.post<any>(postPlanEntrenamiento, planEntrenamiento, {
+      headers: new HttpHeaders({
+        'Content-Type': 'application/json',
+      })
+    });
   }
 }
